Parameterising the error type for RemoteData.
This is a breaking change to the RemoteData API. Previously, all
Failures were assumed to contain Http.Error types. Now they can be any
type.

To convert code to the new API, just change:

type alias Model = {
    people : RemoteData (List Person)
}

...to:

import Http

type alias Model = {
    people : RemoteData Http.Error (List Person)
}

It's a mechanical change, so should be straightforward, but it opens up
the type to deal with other types of remote data that take an
appreciable time to load, like reading data using the HTML5 File API.